Initialize .NET 8.0 C# Scripting. Upgrade to C++20. fmt::format -> std::format.

+++ Source/Runtime/src/Function/Physics/PhysicsErrorCallback.cpp
@@ -7,6 +7,6 @@ namespace SnowLeopardEngine
     PhysicsErrorCallback::reportError(physx::PxErrorCode::Enum code, const char* message, const char* file, int line)
     {
         SNOW_LEOPARD_CORE_ERROR(
-            "[PhysicsSystem] ErrorCode: {0}, Message: {1}, File: {2}, Line: {3}", code, message, file, line);
+            "[PhysicsSystem] ErrorCode: {0}, Message: {1}, File: {2}, Line: {3}", static_cast<int>(code), message, file, line);
     }
 } // namespace SnowLeopardEngine
